feat: refactor code based in PR comments
